extra.views: list people and organisations in group listing
Now that any kind of Contactable (i.e. Person or Organisation) can
belong to a group, when viewing a group listing they must both be
displayed.  The only trouble here is that the specific data like
`name' for orgs and `first_name' for people are located in different
tables, which make the paging and sorting quite difficult.