abook: can now edit people and org contacts
Add extra.forms.ApplicationContactForm for the online
applications and remove some requirements for the standard
ContactForm used in the address book.  Also there is now
ContactForm.is_empty to check whether that contact can be
deleted.

Update extra.views.abook.EditView to handle the ContactForm for
each Contact instance related to the person or org.  Also create
a blank one if the person or organisation has no contact details
at all.  Note that this does not cause anything to be saved if
the contact details remain blank.

Also update templates/abook/edit.html with the contact forms.