abook: add urlopt with sequence of individual URL options
Add `urlopt' as a tuple with 2-tuples containing pairs of option
names and values, which is then used to create the `urlmatch'.
It is also passed on to the template context via the `urlopt'
variable.  This is mainly useful to add these options as hidden
inputs in GET forms, as they can't be passed in the action URL.